Add metadata checks for links and images (#49)
* Add metadata checks for links and images

Metadata checks for links and images have been added to the spam detector. The link check counts the number of URLs in a message and flags it as spam if it exceeds a specified limit. The image check flags messages containing only images as spam. Unit tests for these features have been included.

* add to readme

* Add comments to TelegramListener struct and remove unused fields

* Update README.md to include new meta checks and clarify TG-Spam usage

* change meta-checks execution order to run before len check and allow empty messages with image to processed in listener